refactor: application 패키지 정리
